refactored invites route into controllers
